requirements.txt
flask==3.0.2
tensorflow==2.18.0
torch==2.0.1
torchvision==0.15.2
keras==3.8.0
numpy==1.26.4
opencv-python==4.11.0.86
pillow==11.1.0